+++ src/Components/RequireAuth/RequireAuth.js
@@ -1,9 +1,15 @@
 import { useLocation, Navigate, Outlet } from 'react-router-dom'
 import jwtDecode from 'jwt-decode';
+import Logout from '../../Hooks/Logout'
+
 const RequireAuth = () =>{
     const userToken = localStorage.getItem("zettel_user_token");
     const isTokenActive = jwtDecode(userToken).exp > (new Date().getTime() + 1)/1000
     const location = useLocation();
+    if (!isTokenActive){
+        Logout()
+    }
+
     return (
         isTokenActive 
             ? < Outlet />
